Fix how indicator is added to global-mode-string
We want to pass it as a symbol instead of a string.  This is how other
modes do it as well.  The result is still the same for the user, though
this design ensures that we do not get false positives when we try to
remove our indicator (e.g. if we target an empty string, we can delete
something else).

Thanks to Yusef Aslam for reporting a bug which revealed this problem.
This was done in issue 1 over at the GitHub mirror:
<https://github.com/protesilaos/notmuch-indicator/issues/1>.