Call mapconcat with 3 arguments
Thanks to Stefan Monnier for pointing out that Emacs 27 requires all
three arguments.  What I had before requires Emacs 29.  This
information was provided on the emacs-devel mailing list:
<https://lists.gnu.org/archive/html/emacs-devel/2022-09/msg01949.html>.

The case of mapconcat is the sort of inconsistency that is hard to
detect.  The docstring of the function does not explain as much and
the compiler does not warn about it either.