Make the indicator work over TRAMP
The two new user options, 'notmuch-indicator-notmuch-config-file' and
'notmuch-indicator-notmuch-binary', ensure that we store the details
of the local user at the time the 'notmuch-indicator' feature is
loaded.

We do this because we want the counters to be about the local user's
Notmuch setup: they continue to display the expected values while the
user browses a remote file system with TRAMP.

Before, we did not know how to handle this scenario gracefully: the
counters would be omitted from the mode line if the current buffer was
displaying a file over TRAMP and would be restored once the user
would switch to a local buffer.