Add notmuch-indicator-tab-bar-format
In the past, we were adding the indicator to the global-mode-string,
so it was showing up on the tab-bar via tab-bar-format-global. But now
I prefer to place the indicator in the mode-line-misc-info and make it
easier for users to place it wherever they prefer (I do that with my
own mode-line-format).